pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faq-builder"
version = "0.1.0"
description = ""
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Keith", email = "keith@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "requests",
    "pydantic",
    "click",
    "jinja2",
    "staticjinja",
    "sqlalchemy",
    "python-dateutil",
    "python-dotenv",
    "pyyaml",
    "splink",
    "rapidfuzz",
    "openai>=1.98.0",
    "google-api-python-client",
    "youtube-transcript-api",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
faq-builder = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 150
extend-select = ["I", "F401"]
